pyproject.toml
[project]
name = "ucetni-program-v2"
version = "0.1.0"
description = "Desktopový účetní systém pro české podvojné účetnictví"
requires-python = ">=3.11"
dependencies = [
    "PyQt6>=6.6",
    "weasyprint>=62.0",
    "requests>=2.31",
    "pdfplumber>=0.10.0",
    "pytesseract>=0.3.10",
    "Pillow>=10.0",
    "pdf2image>=1.16.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-qt>=4.4",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["."]

[tool.setuptools.packages.find]
include = ["domain*", "services*", "infrastructure*", "ui*"]
